Removed api.py and moved contents to __main__.py
